refactor: add flow for login page

+++ src/services/article_api.js
@@ -4,12 +4,12 @@ export const articleService = {
     getDataHomepage: (page, size) => api.get('/api/article/homepage', {
         params: {page, size}
     }),
-    getAllArticlesByCategory: (categoryId) => api.get('/api/article/all-articles-by-category', {
+    getAllArticlesByCategory: (categoryId) => api.get('/api/article/admin/all-articles-by-category', {
         params: {categoryId}
     }),
-    getAllDeletedArticles: () => api.get('/api/article/all-deleted-articles'),
-    create: (article) => api.post('/api/article/add', article),
-    update: (articleId, article) => api.put(`/api/article/update/${articleId}`, article),
-    patch: (articleId, articleFields) => api.put(`/api/article/patch/${articleId}`, articleFields),
-    delete: (articleId) => api.delete(`/api/article/delete/${articleId}`),
+    getAllDeletedArticles: () => api.get('/api/article/admin/all-deleted-articles'),
+    create: (article) => api.post('/api/article/admin/add', article),
+    update: (articleId, article) => api.put(`/api/article/admin/update/${articleId}`, article),
+    patch: (articleId, articleFields) => api.put(`/api/article/admin/patch/${articleId}`, articleFields),
+    delete: (articleId) => api.delete(`/api/article/admin/delete/${articleId}`),
 }
